Fix an unintended sign extension
Reported by coverity:
>>>     CID 68353:  Unintended sign extension  (SIGN_EXTENSION)
>>>     Suspicious implicit sign extension: "file_header.e_shnum"
with type "unsigned short" (16 bits, unsigned) is promoted in
"file_header.e_shnum * file_header.e_shentsize" to type "int" (32
bits, signed), then sign-extended to type "unsigned long" (64 bits,
unsigned).  If "file_header.e_shnum * file_header.e_shentsize" is
greater than 0x7FFFFFFF, the upper bits of the result will all be 1.

+++ src/os/linux/host-lib.c
@@ -867,7 +867,7 @@ static int Try_Browser(char *browser, REBCHR *url)
 		return NULL;
 	}
 
-	sec_headers = OS_Make(file_header.e_shnum * file_header.e_shentsize);
+	sec_headers = OS_Make(((size_t)file_header.e_shnum) * file_header.e_shentsize);
 	if (sec_headers == NULL) {
 		fclose(script);
 		return NULL;
